code/requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
scikit-image>=0.21.0
matplotlib>=3.7.0
pyyaml>=6.0
pytest>=7.4.0
scipy>=1.10.0
